Rename helpers/ghostPaths.js -> helpers/ghost-paths.js
Refs #4144
