+++ core/client/app.js
@@ -3,7 +3,7 @@ import loadInitializers from 'ember/load-initializers';
 import 'ghost/utils/link-view';
 import 'ghost/utils/text-field';
 import configureApp from 'ghost/config';
-import ghostPathsHelper from 'ghost/helpers/ghostPaths';
+import ghostPathsHelper from 'ghost/helpers/ghost-paths';
 
 Ember.MODEL_FACTORY_INJECTIONS = true;
