+++ core/client/helpers/ghost-paths.js
